Fix lastmod when using structure bug
Fix bug where the lastmod is always the current date when using structure.

channel_titles data is put in $result, but then the test checks if $date->edit_date is empty, which it always is.

Furthermore, the edit_date in channel_titles is in a stupid format that’s neither usable as a timestamp or a string through localize methods, e.g. 20160127155023, so let's just use some string manipulations to get it as yyyy-mm-dd.